% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAliases.R
\name{contSpectrum}
\alias{contSpectrum}
\title{Continuous Spectrum Alias}
\usage{
contSpectrum(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=getContinuousSpectrum]{getContinuousSpectrum()}}.}
}
\value{
A list with continuous-spectrum results.
}
\description{
Alias for \code{\link[=getContinuousSpectrum]{getContinuousSpectrum()}}.
}
