% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAliases.R
\name{discSpectrum}
\alias{discSpectrum}
\title{Discrete Spectrum Alias}
\usage{
discSpectrum(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=getDiscreteSpectrum]{getDiscreteSpectrum()}}.}
}
\value{
A list with discrete-spectrum results.
}
\description{
Alias for \code{\link[=getDiscreteSpectrum]{getDiscreteSpectrum()}}.
}
