% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParams.R
\name{getParams}
\alias{getParams}
\title{Get ReSpectR Parameters}
\usage{
getParams(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=setParams]{setParams()}}.}
}
\value{
Named list of validated parameters.
}
\description{
Alias for \code{\link[=setParams]{setParams()}} for users who prefer a "get"-style constructor.
}
\examples{
p <- getParams(domain = "frequency", dataFile = "Gst.dat")
p$plateau   # FALSE
}
