% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSpectR-package.R
\docType{package}
\name{ReSpectR-package}
\alias{ReSpectR}
\alias{ReSpectR-package}
\title{ReSpectR: Relaxation Spectrum Extraction in Time and Frequency Domains}
\description{
Tools for continuous and discrete relaxation spectrum extraction from
rheological data in both time-domain G(t) and frequency-domain G*(\eqn{\omega})
formulations.
}
\details{
The algorithm applies Tikhonov regularisation with Bayesian lambda selection
via the L-curve method (continuous spectrum), followed by an AIC-guided
non-negative least-squares search for discrete Maxwell modes.
\subsection{Main workflow}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{par <- setParams(domain = "time", dataFile = "Gt.dat")
crs <- getContinuousSpectrum(par)
drs <- getDiscreteSpectrum(par, crs = crs)
writeSpectrumResults(drs, outputDir = "output")
}\if{html}{\out{</div>}}
}

\subsection{Key functions}{\tabular{ll}{
   Function \tab Purpose \cr
   \code{\link[=setParams]{setParams()}} \tab Build a validated parameter list \cr
   \code{\link[=getContinuousSpectrum]{getContinuousSpectrum()}} \tab Extract continuous H(\eqn{\tau}) spectrum \cr
   \code{\link[=getDiscreteSpectrum]{getDiscreteSpectrum()}} \tab Extract discrete Maxwell modes \cr
   \code{plot()} \tab Plot classed spectrum result objects \cr
   \code{\link[=writeSpectrumResults]{writeSpectrumResults()}} \tab Write results to disk \cr
   \code{\link[=runBatch]{runBatch()}} \tab Process multiple data files \cr
}


Short-form aliases: \code{\link[=contSpectrum]{contSpectrum()}}, \code{\link[=discSpectrum]{discSpectrum()}},
\code{\link[=writeResults]{writeResults()}}, \code{\link[=runTimeBatch]{runTimeBatch()}}, \code{\link[=runFrequencyBatch]{runFrequencyBatch()}}.

Returned objects carry stable classes
\code{"respect_continuous_spectrum"} and \code{"respect_discrete_spectrum"}, so they
can be passed directly to \code{plot()} for built-in visualisation.
}

\subsection{Vignettes}{

The package includes two vignettes:
\itemize{
\item \code{ReSpectR-intro}: end-to-end usage for time- and frequency-domain workflows
\item \code{ReSpectR-python-comparison}: direct R-vs-Python comparison using bundled
reference outputs
}

To open them from an installed package:

\if{html}{\out{<div class="sourceCode">}}\preformatted{browseVignettes("ReSpectR")
}\if{html}{\out{</div>}}
}
}
\references{
Shanbhag, S. (2019). pyReSpect: A computer program to extract discrete and
continuous spectra from stress relaxation experiments. \emph{Macromolecular
Theory and Simulations}, 28(6), 1900005.
}
\seealso{
Package repository: \url{https://github.com/}
}
\author{
\strong{Maintainer}: Giorgio Luciano \email{giorgio.luciano@cnr.it}

Authors:
\itemize{
  \item Giorgio Luciano \email{giorgio.luciano@cnr.it}
}

Other contributors:
\itemize{
  \item Kristian Hovde Liland \email{kristian.liland@nmbu.no} [contributor]
}

}
\keyword{internal}
