% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAliases.R
\name{runTimeBatch}
\alias{runTimeBatch}
\title{Time Batch Alias}
\usage{
runTimeBatch(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=runBatch]{runBatch()}}.}
}
\value{
Data frame with batch status and diagnostics.
}
\description{
Convenience alias for running \code{\link[=runBatch]{runBatch()}} in time domain.
}
