% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiAliases.R
\name{writeResults}
\alias{writeResults}
\title{Write Results Alias}
\usage{
writeResults(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=writeSpectrumResults]{writeSpectrumResults()}}.}
}
\value{
Invisibly returns written file paths.
}
\description{
Alias for \code{\link[=writeSpectrumResults]{writeSpectrumResults()}}.
}
