% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectrum.R
\name{plot_respect_spectrum}
\alias{plot_respect_spectrum}
\alias{plot.respect_continuous_spectrum}
\alias{plot.respect_discrete_spectrum}
\title{Plot ReSpectR Spectrum Results}
\usage{
\method{plot}{respect_continuous_spectrum}(x, y = NULL, ...)

\method{plot}{respect_discrete_spectrum}(x, y = NULL, ...)
}
\arguments{
\item{x}{A result object with class \code{"respect_continuous_spectrum"} or
\code{"respect_discrete_spectrum"}.}

\item{y}{Unused.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Native plotting methods for ReSpectR result objects. Continuous-spectrum
results show the recovered spectrum together with the fitted data curve.
Discrete-spectrum results show Maxwell modes overlaid on the continuous
envelope, plus the fitted response implied by the discrete modes.
}
\examples{
\dontrun{
time_file <- system.file("extdata", "time_tests", "test1.dat", package = "ReSpectR")
par <- setParams(domain = "time", dataFile = time_file, verbose = FALSE)
crs <- getContinuousSpectrum(par)
plot(crs)
drs <- getDiscreteSpectrum(par, crs = crs)
plot(drs)
}
}
